pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mudrex-volume-fees-calculator"
version = "0.1.0"
description = "Calculate API-sourced volume and estimated fees for Mudrex Futures (aligns with mudrex-api-trading-python-sdk)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "mudrex-trading-sdk @ git+https://github.com/DecentralizedJM/mudrex-api-trading-python-sdk.git",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["mudrex_volume_fees*"]

[project.scripts]
mudrex-volume-fees = "mudrex_volume_fees.cli:run"
